# docs/swagger.yaml
basePath: /
definitions:
...
  keerja-backend_internal_utils.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      meta: {}
      success:
        type: boolean
    type: object
host: localhost:3000
info:
  contact:
    email: support@keerja.com
    name: API Support
  description: Job platform backend API with authentication, job management, applications,
    companies, and push notifications
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Keerja Backend API
  version: "1.0"
paths:
...
  /api/v1/meta/locations/cities:
    get:
      consumes:
      - application/json
      description: Retrieve cities with optional filtering by province, active status,
        and search query
      parameters:
      - description: Province ID to filter cities
        in: query
        name: province_id
        required: true
        type: integer
      - description: Filter by active status (true/false)
        in: query
        name: active
        type: boolean
      - description: Search cities by name (min 2 characters)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      summary: Get cities by province
      tags:
      - Master Data - Locations
  /api/v1/meta/locations/cities/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a single city by its ID with province information
      parameters:
      - description: City ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      summary: Get city by ID
      tags:
      - Master Data - Locations
  /api/v1/meta/locations/districts:
    get:
      consumes:
      - application/json
      description: Retrieve districts with optional filtering by city, active status,
        and search query
      parameters:
      - description: City ID to filter districts
        in: query
        name: city_id
        required: true
        type: integer
      - description: Filter by active status (true/false)
        in: query
        name: active
        type: boolean
      - description: Search districts by name (min 2 characters)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      summary: Get districts by city
      tags:
      - Master Data - Locations
  /api/v1/meta/locations/districts/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a single district by its ID with full location hierarchy
        (city and province)
      parameters:
      - description: District ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      summary: Get district by ID
      tags:
      - Master Data - Locations
  /api/v1/meta/locations/provinces:
    get:
      consumes:
      - application/json
      description: Retrieve all provinces with optional filtering by active status
        and search query
      parameters:
      - description: Filter by active status (true/false)
        in: query
        name: active
        type: boolean
      - description: Search provinces by name (min 2 characters)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      summary: Get all provinces
      tags:
      - Master Data - Locations
  /api/v1/meta/locations/provinces/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a single province by its ID
      parameters:
      - description: Province ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/keerja-backend_internal_utils.Response'
      summary: Get province by ID
      tags:
      - Master Data - Locations
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"